// src/codec/codecs/commerce/salesforce-d2c/types.ts
export interface FetchParams {
	url: string,
	method?: 'get' | 'post',
	headers?: {[x: string]: string},
	params?: {[x: string]: string},
	data?: {[x: string]: string}
}

export interface SalesforceD2CCategory {
	id: string;
	name: string;
	slug: string;
	urlName: string;
	children: SalesforceD2CCategory[]
	products: Product[];
	showInMenu: boolean
}

export interface SalesforceD2CMediaItem {
	alternateText: string;
	contentVersionId: null;
	id: string;
	mediaType: string;
	sortOrder: number;
	thumbnailUrl: null;
	title: string;
	url: string;
}

export interface SalesforceD2CMediaGroup {
	developerName: string;
	id: string;
	mediaItems:	SalesforceD2CMediaItem[];
	name: string;
	usageType: string;
}
// ...
export interface SalesforceD2CProductsResp{
	products: SalesforceD2CProductResp[]
}

export interface SalesforceD2CCategoryResp {
	categories: SalesforceD2CCategoryResp[]
	showInMenu: 'true' | 'false'
	bannerImage: {
		alternateText:string;
		contentVersionId: string | null;
		id: string;
		mediaType: string;
		sortOrder: number,
		thumbnailUrl: string;
		title: string;
		url: string;
	},
	fields: {
		MayEdit: 'true' | 'false';
		LastModifiedDate: string;
		IsDeleted: 'true' | 'false';
		Description: string | null;
		LastViewedDate: string;
		UserRecordAccessId: string | null;
		External_ID__c: string;
		IsLocked: 'true' | 'false';
		SortOrder: string;
		LastReferencedDate: string;
		CurrencyIsoCode: string;
		CatalogId: string;
		Name: string;
		SystemModstamp: string;
		CreatedById: string;
		NumberOfProducts: string;
		CreatedDate: string;
		IsNavigational: 'true' | 'false';
		Id: string;
		LastModifiedById: string;
		ParentCategoryId: string | null;
},
	id: string;
	mediaGroups: SalesforceD2CMediaGroup[];
	tileImage: {
		alternateText: string;
		contentVersionId: string | null;
		id: string;
		mediaType: string;
		sortOrder: number;
		thumbnailUrl: string | null;
		title: string;
		url: string;
	},
	urlName: null
}

export interface SalesforceD2CSearchResp {
	categories: {
		category: SalesforceD2CCategory,
		children: SalesforceD2CSearchResp[],
		productCount: 13
	},
	correlationId: string;
	facets: [];
	locale: string;
	productsPage: {
		currencyIsoCode: null;
			pageSize: number;
			products: SalesforceD2CProductResp[];
			total: number;
	}
}

export interface SalesforceD2CCategoriesResp {
	productCategories: SalesforceD2CCategoryResp[];
}

export interface SalesforceD2CAxiosResp {
	data: SalesforceD2CSearchResp & SalesforceD2CCategoryResp & SalesforceD2CProductsResp & SalesforceD2CProductResp & SalesforceD2CCategoriesResp
}